services/anpr-worker/requirements.txt
easyocr==1.7.2
opencv-python-headless==4.10.0.84
Pillow==11.2.1
requests==2.32.3
supabase==2.13.0
python-dotenv==1.1.0
numpy==2.2.6
